Parameter External.5-CallStack

type vt = Val.t

type store_t = Store.t

type t = (string * vt list * store_t option * Gil_syntax.Var.t * int * int * int option) list

Type of call stacks: a call stack is a list of tuples, each of which contains 1) identifier of the current procedure (string) 2) arguments of the current procedure (list of values) 3) calling store 4) variable that should hold the return value 5) index of the calling procedure from whence the procedure was called 6) normal continuation index in the calling procedure 7) optional error continuation index in the calling procedure

val get_cur_proc_id : t -> string

Get current procedure identifier

parameter cs

Target call stack

returns

Identifier of the procedure currently being executed

val get_cur_args : t -> vt list

Get current arguments

parameter cs

Target call stack

returns

Arguments of the procedure currently being executed

val copy : t -> t

Call stack copy

parameter cs

Target call stack

returns

Copy of the given call stack