Module 1-Outcome.ParserAndCompiler

module TargetLangOptions : sig ... end

type err

Type of error that can occur during parsing or compilation

val pp_err : Stdlib.Format.formatter -> err -> unit

Pretty printer for type err

val parse_and_compile_file : string -> ((Gil_syntax.Annot.t, string) Gil_syntax.Prog.t, err) Stdlib.result

Takes a path to a file, parses it with the user's language, and then compiles it to a GIL program. The obtained GIL program should be ready to be analysed.

val other_imports : (string * (string -> ((Gil_syntax.Annot.t, string) Gil_syntax.Prog.t, err) Stdlib.result)) list

other_imports is a association list, that maps extensions to a parser and compiler functions. For example, it is possible to import jsil file in a gil program, using import file.jsil. In order to do so, the other_imports list should contain the double ("jsil", parse_and_compile_jsil_file) where parse_and_compile_jsil_file is a function that takes a file name, parses that jsil file and compiles it to a gil program

val env_var_import_path : string option

Contains the name of the environment variable which contains the path to where the runtime is stored.

val initialize : Utils.ExecMode.t -> unit

Function that will be executed at initialisation. It will be passed the current execution mode as parameter