Module ParserAndCompiler.TargetLangOptions

Target-language specific options

For more help, see Cmdliner documentation

type t

Command line options specific to the target language.

val term : t Cmdliner.Term.t

A term that will be added to every command.

val apply : t -> unit

A side-effect function that will determine the behaviour of the target-language specific options