Module Debugger_lifter

Interface for lifting execution of GIL commands to a target language

type unify_result = Gillian.Debugger.Utils.Unify_map.unify_result =

| Success

| Failure

val unify_result_to_yojson : unify_result -> Yojson.Safe.t

val unify_result_of_yojson : Yojson.Safe.t -> unify_result Ppx_deriving_yojson_runtime.error_or

type ('err, 'annot, 'ast) memory_error_info = {

error : 'err;
(*
The memory error that needs to be lifted
*)

command : (int Gillian.Gil_syntax.Cmd.t * 'annot) option;
(*
The command where it happened
*)

tl_ast : 'ast option;
(*
If the program was compiled from the target language, we keep the tl ast around
*)

}

type 'cmd_report executed_cmd_data = {

kind : (Gillian.Gil_syntax.BranchCase.t, unit) Gillian.Debugger.Utils.Exec_map.cmd_kind;

id : Gillian.Logging.Report_id.t;

cmd_report : 'cmd_report;

unifys : Gillian.Debugger.Utils.Exec_map.unification list;

errors : string list;

branch_path : Gillian.Gil_syntax.BranchCase.path;

}

val executed_cmd_data_to_yojson : 'cmd_report. ('cmd_report -> Yojson.Safe.t) -> 'cmd_report executed_cmd_data -> Yojson.Safe.t

val executed_cmd_data_of_yojson : 'cmd_report. (Yojson.Safe.t -> 'cmd_report Ppx_deriving_yojson_runtime.error_or) -> Yojson.Safe.t -> 'cmd_report executed_cmd_data Ppx_deriving_yojson_runtime.error_or

type handle_cmd_result =

| Stop

| ExecNext of Gillian.Logging.Report_id.t option * Gillian.Gil_syntax.BranchCase.t option

val handle_cmd_result_to_yojson : handle_cmd_result -> Yojson.Safe.t

val handle_cmd_result_of_yojson : Yojson.Safe.t -> handle_cmd_result Ppx_deriving_yojson_runtime.error_or

val _ : Yojson.Safe.t -> handle_cmd_result Ppx_deriving_yojson_runtime.error_or

module type S = sig ... end

val make_executed_cmd_data : (Gillian.Gil_syntax.BranchCase.t, unit) Gillian.Debugger.Utils.Exec_map.cmd_kind -> Gillian.Logging.Report_id.t -> 'cmd_report -> ?unifys:Gillian.Debugger.Utils.Exec_map.unification list -> ?errors:string list -> Gillian.Gil_syntax.BranchCase.path -> 'cmd_report executed_cmd_data

module Gil_lifter : sig ... end

A basic "GIL-to-GIL" lifter implementation.

module Gil_fallback_lifter : sig ... end

A Lifter implementation that acts as a proxy to another lifter (TLLifter), while also keeping a Gil_lifter updated alongside.