Parameter Make.TLLifter

Parameters

module Gil : sig ... end

module V : Engine.Verifier.S with type annot = PC.Annot.t

Signature

type t

The lifter's state

type memory_error = SMemory.err_t

type tl_ast = PC.tl_ast

type memory = SMemory.t

type cmd_report = V.SAInterpreter.Logging.ConfigReport.t

type annot = PC.Annot.t

val init : proc_name:string -> all_procs:string list -> tl_ast option -> (annot, int) Gillian.Gil_syntax.Prog.t -> cmd_report Debugger_lifter__.Lifter_intf.Types.executed_cmd_data -> (t * Debugger_lifter__.Lifter_intf.Types.handle_cmd_result) option

Given a proc name, a tl_ast, and the data from the first executed GIL command, initialise the lifter's state and handle the first command.

Returns None if lifting is unsupported (i.e. if tl_ast is None).

val init_exn : proc_name:string -> all_procs:string list -> tl_ast option -> (annot, int) Gillian.Gil_syntax.Prog.t -> cmd_report Debugger_lifter__.Lifter_intf.Types.executed_cmd_data -> t * Debugger_lifter__.Lifter_intf.Types.handle_cmd_result

Exception-raising version of init.

val dump : t -> Yojson.Safe.t

Gives a JSON representation of the lifter's state.

Used for debugging problems with the lifter.

val handle_cmd : Gillian.Logging.Report_id.t -> Gillian.Gil_syntax.Branch_case.t option -> cmd_report Debugger_lifter__.Lifter_intf.Types.executed_cmd_data -> t -> Debugger_lifter__.Lifter_intf.Types.handle_cmd_result

Handles the execution result of a GIL command.

val get_gil_map : t -> Debugger_utils.Exec_map.Packaged.t

Gets the non-lifted execution map of GIL commands.

In most cases, it's recommended to use a Gil_fallback_lifter, and just defer this call to the GIL lifter.

val get_lifted_map : t -> Debugger_utils.Exec_map.Packaged.t option

Gets the lifted execution map.

Returns None if lifting is not supported.

val get_lifted_map_exn : t -> Debugger_utils.Exec_map.Packaged.t

Exception-raising version of get_lifted_map.

val get_matches_at_id : Gillian.Logging.Report_id.t -> t -> Gillian.Debugger.Utils.Exec_map.matching list

Gives a list of matches that occurred at the specified command.

val get_root_id : t -> Gillian.Logging.Report_id.t option

Gives the id of the root (first) command in the execution map.

val path_of_id : Gillian.Logging.Report_id.t -> t -> Gillian.Gil_syntax.Branch_case.path

Gives the path of (GIL) branch cases that lead to the specified command.

val existing_next_steps : Gillian.Logging.Report_id.t -> t -> (Gillian.Logging.Report_id.t * Gillian.Gil_syntax.Branch_case.t option) list

Gets a list of ID/branch-case pairs that correspond to any commands that directly succeed the specified command.

val next_gil_step : Gillian.Logging.Report_id.t -> Debugger_utils.Exec_map.Packaged.branch_case option -> t -> Gillian.Logging.Report_id.t * Gillian.Gil_syntax.Branch_case.t option

Translates a command ID and (packaged) TL branch case to the command ID and GIL branch case necessary to step forward.

val previous_step : Gillian.Logging.Report_id.t -> t -> (Gillian.Logging.Report_id.t * Debugger_utils.Exec_map.Packaged.branch_case option) option

If the given command has a previous command, gives its ID, and the TL branch case that connects them (if applicable).

val select_next_path : Gillian.Gil_syntax.Branch_case.t option -> Gillian.Logging.Report_id.t -> t -> Gillian.Gil_syntax.Branch_case.path

Gives a branch path that steps forward from the given command.

If the given command branches, the desired branch case can be specified.

val find_unfinished_path : ?at_id:Gillian.Logging.Report_id.t -> t -> (Gillian.Logging.Report_id.t * Gillian.Gil_syntax.Branch_case.t option) option

Gives a branch path under the specified command (or the root command, if not specified) that hasn't yet been fully explored.

The order in which unfinished branches are given in later calls is unimportant.

Returns None if all paths under the specified command have been fully explored.

val memory_error_to_exception_info : (memory_error, annot, tl_ast) Debugger_lifter__.Lifter_intf.Types.memory_error_info -> Gillian.Debugger.Utils.exception_info

val add_variables : store:(string * Gillian.Gil_syntax.Expr.t) list -> memory:memory -> is_gil_file:bool -> get_new_scope_id:(unit -> int) -> Gillian.Debugger.Utils.Variable.ts -> Gillian.Debugger.Utils.Variable.scope list