Module Exec_map.Packaged

An Exec_map to be passed to the debugger frontend and displayed

An Exec_map to be passed to the debugger frontend and displayed

type branch_case = string * Yojson.Safe.t

val branch_case_to_yojson : branch_case -> Yojson.Safe.t

val branch_case_of_yojson : Yojson.Safe.t -> branch_case Ppx_deriving_yojson_runtime.error_or

type t = (branch_case, cmd_data, unit) Exec_map.t

and cmd_data = {

id : Gillian.Logging.Report_id.t;

all_ids : Gillian.Logging.Report_id.t list;

display : string;

unifys : Exec_map.unification list;

errors : string list;

submap : t Exec_map.submap;

}

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val cmd_data_to_yojson : cmd_data -> Yojson.Safe.t

val cmd_data_of_yojson : Yojson.Safe.t -> cmd_data Ppx_deriving_yojson_runtime.error_or

val _ : Yojson.Safe.t -> cmd_data Ppx_deriving_yojson_runtime.error_or

val package_gil_case : Gillian.Gil_syntax.Branch_case.t -> branch_case

Converts a GIL branch case to a packaged branch case

val package : ((('c, 'd, 'bd) Exec_map.t -> t) -> 'd0 -> cmd_data) -> (bd:'bd0 -> all_cases:('c0 * 'bd1) list -> 'c1 -> branch_case) -> ('c, 'd, 'bd) Exec_map.t -> t

Converts an Exec_map to a packaged Exec_map