Module type Unify_map.Build

Build.f: Given the ID of a unification, build the representative unification map

val f : Gillian.Logging.Report_id.t -> Engine.Unifier.unify_kind * Debugger_utils__.Unify_map_intf.Types.map

Given the ID of a unification, build the representative unification map