Module Engine.UP

type outs = (Gil_syntax.Expr.t * Gil_syntax.Expr.t) list

val outs_pp : outs Fmt.t

type step = Gil_syntax.Asrt.t * outs

The up_step type represents a unification plan step, consisting of an assertion together with the possible learned outs

val step_to_yojson : step -> Yojson.Safe.t

val step_of_yojson : Yojson.Safe.t -> step Ppx_deriving_yojson_runtime.error_or

val step_pp : step Fmt.t

type t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

type pred = {

pred : Gil_syntax.Pred.t;

up : t;

}

type spec = {

spec : Gil_syntax.Spec.t;

up : t;

}

type lemma = {

lemma : Gil_syntax.Lemma.t;

up : t;

}

type 'annot prog = {

preds : ( string, pred ) Utils.Prelude.Hashtbl.t;

specs : ( string, spec ) Utils.Prelude.Hashtbl.t;

lemmas : ( string, lemma ) Utils.Prelude.Hashtbl.t;

coverage : ( string * int, int ) Utils.Prelude.Hashtbl.t;

prog : ( 'annot, int ) Gil_syntax.Prog.t;

}

type preds_tbl_t = ( string, pred ) Utils.Prelude.Hashtbl.t

type up_err_t =

| UPSpec of string * Gil_syntax.Asrt.t list list

| UPPred of string * Gil_syntax.Asrt.t list list

| UPLemma of string * Gil_syntax.Asrt.t list list

| UPAssert of Gil_syntax.Asrt.t * Gil_syntax.Asrt.t list list

| UPInvariant of Gil_syntax.Asrt.t * Gil_syntax.Asrt.t list list

val pp_up_err_t : Ppx_deriving_runtime.Format.formatter -> up_err_t -> Ppx_deriving_runtime.unit

val show_up_err_t : up_err_t -> Ppx_deriving_runtime.string

module KB = Gil_syntax.Expr.Set

val learn_expr : ?top_level:bool -> KB.t -> Gil_syntax.Expr.t -> Gil_syntax.Expr.t -> outs

val ins_outs_expr : KB.t -> Gil_syntax.Expr.t -> Gil_syntax.Expr.t -> (KB.t * outs) list

val collect_simple_asrts : Gil_syntax.Asrt.t -> Gil_syntax.Asrt.t list

val empty_up : t

val init : ?use_params:bool -> KB.t -> KB.t -> ( string, int list ) Utils.Prelude.Hashtbl.t -> (Gil_syntax.Asrt.t * ((string * Utils.Containers.SS.t) option * (Gil_syntax.Flag.t * Gil_syntax.Asrt.t list) option * string list)) list -> ( t, Gil_syntax.Asrt.t list list ) Stdlib.result

val next : t -> (t * (string * Utils.Containers.SS.t) option) list option

val head : t -> step option

val posts : t -> (Gil_syntax.Flag.t * Gil_syntax.Asrt.t list) option

val hides : t -> string list option

val init_prog : ?preds_tbl:( string, pred ) Utils.Prelude.Hashtbl.t -> ( 'a, int ) Gil_syntax.Prog.t -> ( 'a prog, up_err_t ) Stdlib.result

val init_preds : ( string, Gil_syntax.Pred.t ) Utils.Prelude.Hashtbl.t -> ( ( string, pred ) Utils.Prelude.Hashtbl.t, up_err_t ) Stdlib.result

val pp : Stdlib.Format.formatter -> t -> unit

val get_pred_def : preds_tbl_t -> string -> pred

val init_pred_defs : unit -> preds_tbl_t

val pp_pred_defs : Stdlib.Format.formatter -> preds_tbl_t -> unit

val get_procs : 'a prog -> ( 'a, int ) Gil_syntax.Proc.t list

val get_bispecs : 'a prog -> Gil_syntax.BiSpec.t list

val pp_asrt : ?preds_printer: ( Stdlib.Format.formatter -> (string * Gil_syntax.Expr.t list) -> unit ) -> preds:preds_tbl_t -> Stdlib.Format.formatter -> Gil_syntax.Asrt.t -> unit

val pp_spec : ?preds_printer: ( Stdlib.Format.formatter -> (string * Gil_syntax.Expr.t list) -> unit ) -> preds:preds_tbl_t -> Stdlib.Format.formatter -> Gil_syntax.Spec.t -> unit

val pp_normal_spec : ?preds_printer: ( Stdlib.Format.formatter -> (string * Gil_syntax.Expr.t list) -> unit ) -> preds:preds_tbl_t -> Stdlib.Format.formatter -> Gil_syntax.Spec.t -> unit

val add_spec : 'a prog -> Gil_syntax.Spec.t -> unit

val remove_spec : 'a prog -> string -> unit

val get_lemma : 'a prog -> string -> ( lemma, unit ) Stdlib.result

val update_coverage : 'a prog -> string -> int -> unit

val first_time_running : 'a prog -> string -> int -> bool