Module Engine__PFS

type t

Type of Gillian pure formulae

val init : unit -> t

init () returns a fresh empty collection of pure formulae

val equal : t -> t -> bool

equal pfs_a pfs_b returns true iff pfs_a = pfs_b

val to_list : t -> Gil_syntax.Formula.t list

to_list pfs serialises the pure formulae pfs into a list

val of_list : Gil_syntax.Formula.t list -> t

of_list fs deserialises a list of formulae fs into pure formulae

val mem : t -> Gil_syntax.Formula.t -> bool

mem pfs f return true iff the formula f is part of the pure formulae pfs

val extend : t -> Gil_syntax.Formula.t -> unit

extend pfs f extends the pure formulae pfs with the formula f

val nth_get : t -> int -> Gil_syntax.Formula.t

nth_get pfs n returns the n-th pure formula of pfs

val nth_set : t -> int -> Gil_syntax.Formula.t -> unit

nth_set pfs n f sets the n-th pure formula of pfs to equal f

val nth_delete : t -> int -> unit

nth_delete pfs n f deletes the n-th pure formula of pfs

val clear : t -> unit

clear pfs empties the pure formulae pfs

val length : t -> int

length pfs returns the length of the pure formulae pfs

val copy : t -> t

copy pfs Returns a copy of the pure formulae pfs

val merge_into_left : t -> t -> unit

merge_into_left pfsl pfsr merges the pure formulae pfsr with the pure formulae pfsl

val set : t -> Gil_syntax.Formula.t list -> unit

set pfs fs sets the pure formulae pfs to fs

val iter : (Gil_syntax.Formula.t -> unit) -> t -> unit

iter f pfs iterates over the pure formulae pfs using the function f

val iteri : (int -> Gil_syntax.Formula.t -> unit) -> t -> unit

iteri f pfs iterates over the pure formulae pfs using the function f, keeping track of the index

val fold_left : ('a -> Gil_syntax.Formula.t -> 'a) -> 'a -> t -> 'a

fold_left f ac pfs folds over the pure formulae pfs using the function f and initial accumulator ac

val substitution : Engine.SVal.SSubst.t -> t -> unit

substitution subst pfs substitutes the substutition subst in the pure formulae pfs in-place

val subst_expr_for_expr : Gil_syntax.Expr.t -> Gil_syntax.Expr.t -> t -> unit

subst_expr_for_expr e_to_subst e_subst pfs substitutes the expression e_to_subst with the expression e_subst in the pure formulae pfs in-place

val lvars : t -> Utils.Containers.SS.t

lvars pfs returns the set containing all the lvars occurring in pfs

val alocs : t -> Utils.Containers.SS.t

Returns the set containing all the alocs occurring in --pfs--

alocs pfs returns the set containing all the abstract locations occurring in pfs

val clocs : t -> Utils.Containers.SS.t

clocs pfs returns the set containing all the concrete locations occurring in pfs

val count_lvar : t -> Gil_syntax.Var.t -> int

count_lvar pfs x returns the number of formulae in which the variable x occurs

val pp : Stdlib.Format.formatter -> t -> unit

pp fmt pfs prints the pure formulae pfs

val sort : t -> unit

sort pfs sorts the pure formulae pfs