Module Gil_parsing

module Preprocess_GCmd : sig ... end

val log_verboser : ((('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a) -> unit) -> unit

val col : Stdlib.Lexing.position -> int

val parse : ((Stdlib.Lexing.lexbuf -> Gil_parsing__.GIL_Parser.token) -> Stdlib.Lexing.lexbuf -> 'a) -> Stdlib.Lexing.lexbuf -> 'a

val parse_from_string : ((Stdlib.Lexing.lexbuf -> Gil_parsing__.GIL_Parser.token) -> Stdlib.Lexing.lexbuf -> 'a) -> string -> 'a

val parse_eprog_from_string : string -> (Gil_syntax.Annot.t, string) Gil_syntax.Prog.t

val parse_expr_from_string : string -> Gil_syntax.Expr.t

val parse_eprog_from_file : string -> (Gil_syntax.Annot.t, string) Gil_syntax.Prog.t

val eprog_to_prog : other_imports:(string * (string -> (Gil_syntax.Annot.t, string) Gil_syntax.Prog.t)) list -> (Gil_syntax.Annot.t, string) Gil_syntax.Prog.t -> (Gil_syntax.Annot.t, int) Gil_syntax.Prog.t

Converts a string Prog.t to a Prog.t. Resolving the imports in the meantime. The parameter other_imports is a association list, that maps extensions to a parser and compiler. For example, it is possible to import jsil file in a gil program, using import file.jsil. In order to do so, the other_imports list should contain the double ("jsil", parse_and_compile_jsil_file) where parse_and_compile_jsil_file is a function that takes a file name, parses that jsil file and compiles it to a gil program.

val parse_line_numbers : string -> (string * int, int * bool) Stdlib.Hashtbl.t

---------------------------------------------------- Parse a line_numbers file. Proc: proc_name (0, 0) ... -----------------------------------------------------