Module Gil_syntax.Expr

type t =

| Lit of Literal.t	(*
GIL literals
*)

| PVar of string	(*
GIL program variables
*)

| LVar of string	(*
GIL logical variables (interpreted symbols)
*)

| ALoc of string	(*
GIL abstract locations (uninterpreted symbols)
*)

| UnOp of UnOp.t * t	(*
Unary operators (UnOp.t)
*)

| BinOp of t * BinOp.t * t	(*
Binary operators (BinOp.t)
*)

| LstSub of t * t * t	(*
Sublist
*)

| NOp of NOp.t * t list	(*
n-ary operators (NOp.t)
*)

| EList of t list	(*
Lists of expressions
*)

| ESet of t list	(*
Sets of expressions
*)

GIL Expressions

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val lit : Literal.t -> t

val num : float -> t

val num_int : int -> t

val int : int -> t

val int_z : Z.t -> t

val string : string -> t

val bool : bool -> t

val zero_i : t

Lit (Int Z.zero)

val one_i : t

Lit (Int Z.one)

val type_ : Type.t -> t

val list : t list -> t

val list_length : t -> t

val list_nth : t -> int -> t

val list_nth_e : t -> t -> t

val list_sub : lst:t -> start:t -> size:t -> t

val list_cons : t -> t -> t

val list_cat : t -> t -> t

val typeof : t -> t

val fmod : t -> t -> t

val imod : t -> t -> t

module Infix : sig ... end

module Set : Stdlib.Set.S with type elt := t

Sets of expressions

module Map : Utils.Prelude.Map.S with type key := t

Maps with expressions as keys

val equal : t -> t -> bool

Equality

Mapper

val map_opt : ( t -> t option * bool ) -> ( t -> t ) option -> t -> t option

Optional mapper

val pp : t Fmt.t

Pretty-printer

val full_pp : t Fmt.t

Pretty-printer with constructors (will not parse)

val to_list : t -> t list option

If the expression is a list (either an EList _ of Lit (LList _)), returns the list of expressions.

val from_list : t list -> t

from_list EList with the provided elements

val lvars : t -> Utils.Prelude.SS.t

lvars e returns all logical variables in e

val pvars : t -> Utils.Prelude.SS.t

pvars e returns all program variables in e

val alocs : t -> Utils.Prelude.SS.t

alocs e returns all abstract locations in e

val clocs : t -> Utils.Prelude.SS.t

clocs e returns all concrete locations in e

val locs : t -> Utils.Prelude.SS.t

locs e returns all concrete and abstract locations in e

val vars : t -> Utils.Prelude.SS.t

vars e returns all variables in e (includes lvars, pvars, alocs and clocs)

val substitutables : t -> Utils.Prelude.SS.t

substitutables e returns all lvars and alocs

val is_concrete : t -> bool

is_concrete e returns true iff the expression contains no lvar or aloc

val all_literals : t list -> bool

all_literals lst returns true iff all elements of the given list lst are literals

val from_lit_list : Literal.t -> t

from_lit_list lst lifts a literal list to an expression list

val lists : t -> t list

lists e all sub-expressions of e of the form Lit (LList lst) and EList lst

val subst_clocs : ( string -> t ) -> t -> t

subst_clocs subst e substitutes expressions of the form Lit (Loc l) with subst l in e

val from_var_name : string -> t

from_var_name var returns either an aloc, an lvar or a pvar if var name matches one of these types (see Utils.Names.is_aloc_name, Utils.Names.is_lvar_name and Utils.Names.is_pvar_name)

val loc_from_loc_name : string -> t

loc_from_loc_name loc Has the same behaviour as from_var_name except that it returns either an ALoc loc or a Lit (Loc loc)

val subst_expr_for_expr : to_subst:t -> subst_with:t -> t -> t

subst_expr_for_expr ~to_subst ~subst_with expr substitutes every occurence of the expression to_subst with the expression subst_with in expr

val base_elements : t -> t list

base_elements e returns the list containing all logical variables, abstract locations, and non-list literals in e

val var_to_expr : string -> t

var_to_expr x returns the expression representing the program/logical variable or abstract location x

val is_unifiable : t -> bool

is_unifiable x returns whether or not the expression e is unifiable

val sub_expr : t -> t -> bool

Sub-expression check