Module Debugger.Utils

Miscellaneous types and functions for debugger-related purposes

val location_to_display_location : Gil_syntax.Location.t -> Gil_syntax.Location.t

Converts a GIL location (column number starts from 0) to a VSCode location (column number starts from 1)

type stop_reason =

| Step
(*
The step has been completed
*)

| ReachedStart
(*
The start of the program has been reached (when executing backwards)
*)

| ReachedEnd
(*
The end of the program has been reached
*)

| Breakpoint
(*
A breakpoint has been reached
*)

| ExecutionError
(*
An error in execution has occurred
*)

Explains why the debugger has stopped

val stop_reason_to_yojson : stop_reason -> Yojson.Safe.t

val stop_reason_of_yojson : Yojson.Safe.t -> stop_reason Ppx_deriving_yojson_runtime.error_or

val _ : Yojson.Safe.t -> stop_reason Ppx_deriving_yojson_runtime.error_or

type frame = {

index : int;
(*
The index of the frame in the stack
*)

name : string;
(*
The name of the frame
*)

source_path : string;
(*
The path to the source file of the relevant code
*)

start_line : int;
(*
The line number of the start of the relevant code
*)

start_column : int;
(*
The column number of the start of the relevant code
*)

end_line : int;
(*
The line number of the end of the relevant code
*)

end_column : int;
(*
The column number of the end of the relevant code
*)

}

Describes a frame on the stack of execution

val make_frame : index:int -> name:string -> source_path:string -> start_line:int -> start_column:int -> end_line:int -> end_column:int -> frame

type exception_info = {

id : string;

description : string option;

}

Describes an exception

module Variable : sig ... end

Representation of variables for the debugger

module Exec_map : sig ... end

A map of execution of a function (symbolic or otherwise).

module Match_map : sig ... end

A map of a matching.