Module Exec_map.Packaged

An Exec_map to be passed to the debugger frontend and displayed

type branch_case = {

kind : string;

display : string * string;	(*
A friendly name for the branch kind and specific branch case to be displayed to the user
*)

json : Yojson.Safe.t;	(*
The JSON of the original branch case; this can be target language specific
*)

}

val branch_case_to_yojson : branch_case -> Yojson.Safe.t

val branch_case_of_yojson : Yojson.Safe.t -> branch_case Ppx_deriving_yojson_runtime.error_or

type t = ( branch_case, cmd_data, unit ) t

and cmd_data = {

ids : Logging.Report_id.t list;

display : string;

unifys : unification list;

errors : string list;

submap : t submap;

}

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val cmd_data_to_yojson : cmd_data -> Yojson.Safe.t

val cmd_data_of_yojson : Yojson.Safe.t -> cmd_data Ppx_deriving_yojson_runtime.error_or

val _ : Yojson.Safe.t -> cmd_data Ppx_deriving_yojson_runtime.error_or

val package_gil_case : Gil_syntax.BranchCase.t -> branch_case

Converts a GIL branch case to a packaged branch case

val package : ( ( ( 'c, 'd, 'bd ) t -> t ) -> 'd0 -> cmd_data ) -> ( 'bd0 -> 'c0 -> branch_case ) -> ( 'c, 'd, 'bd ) t -> t

Converts an Exec_map to a packaged Exec_map