Module type Match_map.Build

Build.f: Given the ID of a matching, build the representative matching map

val f : ?pp_asrt:Gil_syntax.Asrt.atom Fmt.t -> Logging.Report_id.t -> Debugger_utils__.Match_map_intf.Types.t

Given the ID of a matching, build the representative matching map