Module Utils.Unify_map

A map of a unification.

A unification is either a single string of assertions for most unifications (Direct) or potentially multiple strings of assertions when folding a predicate (Fold), each representing a case of the predicate.

Each string of assertions (dubbed unify_seg) is essentially a singly linked list of Assertions (each of which contains information about that assertion - see assertion_data), terminated by a UnifyResult that states whether that unification (or that case in the unification) was a success.

type kind = Engine.Unifier.unify_kind

Describes why this unification is happening

val kind_to_yojson : kind -> Yojson.Safe.t

val kind_of_yojson : Yojson.Safe.t -> kind Ppx_deriving_yojson_runtime.error_or

type unify_result =

| Success

| Failure

val unify_result_to_yojson : unify_result -> Yojson.Safe.t

val unify_result_of_yojson : Yojson.Safe.t -> unify_result Ppx_deriving_yojson_runtime.error_or

type substitution = {

assert_id : Logging.Report_id.t;

subst : string * string;

}

A substitution, and the ID of the assertion where it was learned

val substitution_to_yojson : substitution -> Yojson.Safe.t

val substitution_of_yojson : Yojson.Safe.t -> substitution Ppx_deriving_yojson_runtime.error_or

type assertion_data = {

id : Logging.Report_id.t;	(*
The report ID of the assertion in the log database
*)

fold : (Logging.Report_id.t * unify_result) option;	(*
The ID of the fold unification and its result, if this assertion requires a fold
*)

assertion : string;	(*
The string representation of this assertion
*)

substitutions : substitution list;	(*
A list of the substitutions learned from this assertion specifically
*)

}

Represents one step of a unification

val assertion_data_to_yojson : assertion_data -> Yojson.Safe.t

val assertion_data_of_yojson : Yojson.Safe.t -> assertion_data Ppx_deriving_yojson_runtime.error_or

type unify_seg =

| Assertion of assertion_data * unify_seg	(*
A single assertion
*)

| UnifyResult of Logging.Report_id.t * unify_result	(*
The end of this unification segment
*)

A segment of unification

val unify_seg_to_yojson : unify_seg -> Yojson.Safe.t

val unify_seg_of_yojson : Yojson.Safe.t -> unify_seg Ppx_deriving_yojson_runtime.error_or

type map =

| Direct of unify_seg

| Fold of unify_seg list

A unification map. Unification is either a single segment in the normal case, or potentially multiple segments when folding (i.e. when a predicate has multiple cases)

val map_to_yojson : map -> Yojson.Safe.t

val map_of_yojson : Yojson.Safe.t -> map Ppx_deriving_yojson_runtime.error_or

type t = kind * map

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val _ : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val result : t -> unify_result

module type Build = sig ... end

Build.f: Given the ID of a unification, build the representative unification map

module Make_builder : sig ... end