Parameter Make.Val

type t

Type of GIL values

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

type st

Type of substitutions for GIL values

type et

Type of extended substitutions for GIL values

val equal : t -> t -> bool

val pp : Stdlib.Format.formatter -> t -> unit

Printer

val full_pp : Stdlib.Format.formatter -> t -> unit

val full_pp_list : Stdlib.Format.formatter -> t list -> unit

val to_literal : t -> Gil_syntax.Literal.t option

Convert a value to a literal, if possible

val from_literal : Gil_syntax.Literal.t -> t

Convert a literal to a value, always possible

val to_expr : t -> Gil_syntax.Expr.t

Convert a value to a logical expression, always possible

val from_expr : Gil_syntax.Expr.t -> t option

Converts a logical expression to a value, if possible

val from_list : t list -> t

Convert a list of values to a single value, always possible

val from_lvar_name : string -> t

Converts a logical variable name into a value

val to_list : t -> t list option

Convert a value to a list of values, if possible