Module Logging.Phase

Functions for managing phases

val normal : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> unit -> Logging.Report_id.t option

Starts a phase with logging level set to Normal

val verbose : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> unit -> Logging.Report_id.t option

Starts a phase with logging level set to Verbose

val tmi : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> unit -> Logging.Report_id.t option

Starts a phase with logging level set to TMI

val stop : Logging.Report_id.t option -> unit

Ends the phase corresponding to the specified report id

val with_normal : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> (unit -> 'a) -> 'a

Runs the specified function within a phase with logging level set to Normal

val with_verbose : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> (unit -> 'a) -> 'a

Runs the specified function within a phase with logging level set to Verbose

val with_tmi : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> (unit -> 'a) -> 'a

Runs the specified function within a phase with logging level set to TMI