Module Logging.Specific

Logs a type given a Loggable.t and its content type (which should be one of the predefined strings in Logging_constants.Content_type). Returns the logged report id if it has been logged.

val normal : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> Loggable.t -> string -> Logging.Report_id.t option

Logs a Loggable.t at the Normal log level

val verbose : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> Loggable.t -> string -> Logging.Report_id.t option

Logs a Loggable.t at the Verbose log level

val tmi : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> Loggable.t -> string -> Logging.Report_id.t option

Logs a Loggable.t at the TMI log level