Module Utils.Allocators

Allocators

val register_resetter : (unit -> unit) -> unit

Allocators's resetters should be registers so that Bulk Execution can reset them at every start

val reset_all : unit -> unit

module type S = sig ... end

module type S_with_stringify = sig ... end

module Basic : functor () S_with_stringify with type t = int

Creating a BasicIdentifier module also registers its resetter on its own

module Make_with_prefix : functor (A : S_with_stringify) -> functor (P : sig ... end) -> S with type t = string

Make_with_prefix is assuming that one already registered the resetter of A