Module Utils.List_res

type ('a, 'b) t = ('a list, 'b list) Stdlib.result

This module implements what could be called the "verification" monad. Each step may branch, but if any branch fails, then the entirety of the process fails.

val return : 'a -> ('a0, 'b) t

val bind : ('a, 'e) t -> ('a0 -> ('b, 'e0) t) -> ('b0, 'e1) t

val map : ('a -> 'b) -> ('a0, 'e) t -> ('b0, 'e0) t

module Syntax : sig ... end