Module Utils__Config

This file contains some configurations for the framework. Mostly composed of modifiable flags, or hardcoded values.

val ci : bool Stdlib.ref

global config

val current_exec_mode : Utils.ExecMode.t Stdlib.ref

val previously_normalised : bool Stdlib.ref

Hashtable sizes

val small_tbl_size : int

val medium_tbl_size : int

val big_tbl_size : int

Bi-abduction configuration

These values seem to never be modified..

val bug_specs_propagation : bool Stdlib.ref

This value seems to never be modified

val output_verification : bool Stdlib.ref

This value seems to never be modified

Debugging configuration

val debug : bool Stdlib.ref

Printing configuration

val no_heap : bool Stdlib.ref

Global configuration

val unfolding : bool Stdlib.ref

val manual_proof : bool Stdlib.ref

val lemma_proof : bool Stdlib.ref

Parallel threading

val multi_thread : bool Stdlib.ref

val parallel : bool Stdlib.ref

Statistics

val stats : bool Stdlib.ref

Symbolic execution

val bi_dflt : bool Stdlib.ref

val bi_unfold_depth : int Stdlib.ref

val delay_entailment : bool Stdlib.ref

Bulk testing

val bulk_print_all_failures : bool Stdlib.ref

If activated, at the end of bulk execution or bulk wpst, and only with the Rely runner, a list of all failures will be printed in stdout

val set_runtime_paths : ?⁠env_var:string -> string list -> unit

Runtime settings

val get_runtime_paths : unit -> string list

Runtime settings