Module type Utils__Union_find.S

type elt

type t

val init : unit -> t

init () initialises a new, empty union-find

val copy : t -> t

copy uf copies the given union-find

val add : t -> elt -> unit

add uf e creates a new element e in the union-find uf

val rep : t -> elt -> elt

rep uf e finds the representative of node e in the union-find uf

val union : t -> elt -> elt -> unit

union uf e1 e2 merges elements e1 and e2 in the union-find uf, performing path compression along the way.

val fold : ('a -> (elt * elt) -> 'a) -> 'a -> t -> 'a

fold f ac uf folds over the union find uf