Module Debugger_utils.Exec_map

A map of execution of a function (symbolic or otherwise).

This is represented in a tree structure, where each node represents a command.

Each command node comes in one of these forms:

Normal ( Cmd ) - a normal command with one command node as its child.

Branching (BranchCmd) - a command with one or more children, with each child identified by a branch case type; this represents a command that has multiple potential outcomes, depending on the program state (e.g. an if/else). Note that this is only interesting in symbolic execution; in concrete execution, they will always have exactly one branch case, thus functioning similarly to normal command nodes.

Final (FinalCmd) - a command with no children; this represents the end of an execution branch, either due to errors, or normal termination (i.e. when returning from the function).

These nodes also contain some (configurable) set of information - this usually includes the relevant report ID from the log database, a human-readable representation of the command, and any errors and unifications that occurred during executing the command.

The child of a Cmd node may be Nothing, as can any child of a BranchCmd node; this represents that there is a command here in the code, but it hasn't yet been executed. This is to facilitate the debugger's step-by-step behaviour.

A command node may also contain a submap; this embeds another exec map inside this command node, either described in full or referred to by name (see submap). This is used to, for example, embed the body of a while-loop in the while-loop command itself.

type ('c, 'bd) cmd_kind =

| Normal	(*
A command that doesn't branch or terminate
*)

| Branch of ('c * 'bd) list	(*
A branching command
*)

| Final	(*
A terminating command
*)

The "kind" of a command in an exec map

val cmd_kind_to_yojson : 'c 'bd. ( 'c -> Yojson.Safe.t ) -> ( 'bd -> Yojson.Safe.t ) -> ( 'c, 'bd ) cmd_kind -> Yojson.Safe.t

val cmd_kind_of_yojson : 'c 'bd. ( Yojson.Safe.t -> 'c Ppx_deriving_yojson_runtime.error_or ) -> ( Yojson.Safe.t -> 'bd Ppx_deriving_yojson_runtime.error_or ) -> Yojson.Safe.t -> ( 'c, 'bd ) cmd_kind Ppx_deriving_yojson_runtime.error_or

val kind_of_cases : 'a list -> ( 'b, unit ) cmd_kind

Maps a list of branches to Normal if empty, or Branch

type ('branch_case, 'cmd_data, 'branch_data) t =

| Nothing	(*
An empty space; represents a command yet to be executed
*)

| Cmd of {

data : 'cmd_data;

mutable next : ( 'branch_case, 'cmd_data, 'branch_data ) t;

}	(*
A non-branching command with one next command
*)

| BranchCmd of {

data : 'cmd_data;

nexts : ( 'branch_case, 'branch_data * ( 'branch_case, 'cmd_data, 'branch_data ) t ) Utils.Prelude.Hashtbl.t;

}	(*
A branching command, with one or more branch cases
*)

| FinalCmd of {

data : 'cmd_data;

}	(*
A command with no subsequent ones, either due to normal termination or an error
*)

An exec map / node in an exec map; takes the following type parameters:

'branch_case: the type that identifies a branch case

'cmd_data: the type of the data attached to each non-Nothing node

'branch_data: additional data attached to each branch case

val to_yojson : 'branch_case 'cmd_data 'branch_data. ( 'branch_case -> Yojson.Safe.t ) -> ( 'cmd_data -> Yojson.Safe.t ) -> ( 'branch_data -> Yojson.Safe.t ) -> ( 'branch_case, 'cmd_data, 'branch_data ) t -> Yojson.Safe.t

val of_yojson : 'branch_case 'cmd_data 'branch_data. ( Yojson.Safe.t -> 'branch_case Ppx_deriving_yojson_runtime.error_or ) -> ( Yojson.Safe.t -> 'cmd_data Ppx_deriving_yojson_runtime.error_or ) -> ( Yojson.Safe.t -> 'branch_data Ppx_deriving_yojson_runtime.error_or ) -> Yojson.Safe.t -> ( 'branch_case, 'cmd_data, 'branch_data ) t Ppx_deriving_yojson_runtime.error_or

type stop_at =

| StartOfPath	(*
Stop at the start of the path, i.e. as soon as possible
*)

| EndOfPath	(*
Stop at the end of the path, i.e. as late as possible
*)

| BeforeNothing	(*
As with EndOfPath, but if the path ends with Nothing, step back to the previous command
*)

Used for various map-traversal commands; signifies when to stop exploring paths

type unification = {

id : Logging.Report_id.t;

kind : Engine.Unifier.unify_kind;

result : Unify_map.unify_result;

}

Data about a unification

val unification_to_yojson : unification -> Yojson.Safe.t

val unification_of_yojson : Yojson.Safe.t -> unification Ppx_deriving_yojson_runtime.error_or

type 't submap =

| NoSubmap

| Submap of 't	(*
Embed an Exec_map as a submap
*)

| Proc of string	(*
Embed the execution of another proc as a submap
*)

val submap_to_yojson : 't. ( 't -> Yojson.Safe.t ) -> 't submap -> Yojson.Safe.t

val submap_of_yojson : 't. ( Yojson.Safe.t -> 't Ppx_deriving_yojson_runtime.error_or ) -> Yojson.Safe.t -> 't submap Ppx_deriving_yojson_runtime.error_or

val _ : ( Yojson.Safe.t -> 'a Ppx_deriving_yojson_runtime.error_or ) -> Yojson.Safe.t -> 'a submap Ppx_deriving_yojson_runtime.error_or

val find_path : ( 'a -> bool ) -> ( 'b, 'c, 'd ) t -> 'b list option

Traverse the map depth-first, giving the path to the first node that matches the given predicate (or None otherwise)

val find_path_exn : ( 'a -> bool ) -> ( 'b, 'a, 'c ) t -> 'b list

Exception-raising equivalent to find_path

val at_path : ?stop_at:stop_at -> 'a list -> ( 'b, 'c, 'd ) t -> ( 'b, 'c, 'd ) t option

Gets the node at the given path

val at_path_exn : ?stop_at:stop_at -> 'a list -> ( 'a, 'b, 'c ) t -> ( 'a, 'b, 'c ) t

Exception-raising equivalent to at_path

module Packaged : sig ... end

An Exec_map to be passed to the debugger frontend and displayed