Module Gil_syntax.Asrt

type t =

| Emp
Empty heap

| Star of t * t
Separating conjunction

| Pred of string * Expr.t list
Predicates

| Pure of Formula.t
Pure formula

| Types of (Expr.t * Type.t) list
Typing assertion

| GA of string * Expr.t list * Expr.t list
Core assertion

GIL Assertions

val compare : t -> t -> int

Comparison of assertions

val prioritise : t -> t -> int

Sorting of assertions

module Set : Stdlib.Set.S with type Set.elt := t

Sets of assertions

val map : (t -> t * bool) option -> (t -> t) option -> (Expr.t -> Expr.t) option -> (Formula.t -> Formula.t) option -> t -> t

Deprecated, use Visitors.map instead.

val fold : (Expr.t -> 'a) option -> (Formula.t -> 'a) option -> (t -> 'b -> 'b -> 'a list -> 'a) -> (t -> 'b -> 'b) option -> 'b -> t -> 'a

val lists : t -> Expr.t list

Get all the logical expressions of a of the form (Lit (LList lst)) and (EList lst)

val list_lexprs : t -> Expr.t list

Get all the logical expressions of a that denote a list and are not logical variables

val lvars : t -> Utils.Containers.SS.t

Get all the logical variables in a

val pvars : t -> Utils.Containers.SS.t

Get all the program variables in a

val alocs : t -> Utils.Containers.SS.t

Get all the abstract locations in a

val clocs : t -> Utils.Containers.SS.t

Get all the concrete locations in a

val vars : t -> Utils.Containers.SS.t

Get all the variables in a (includes lvars, pvars, alocs and clocs)

val pred_names : t -> string list

Returns a list with the names of the predicates that occur in a

val pure_asrts : t -> Formula.t list

Returns a list with the pure assertions that occur in a

val simple_asrts : t -> t list

Returns a list with the pure assertions that occur in a

val is_pure_asrt : t -> bool

Check if a is a pure assertion

val is_pure_non_rec_asrt : t -> bool

Check if a is a pure assertion & non-recursive assertion. It assumes that only pure assertions are universally quantified

val make_pure : t -> Formula.t

Eliminate LStar and LTypes assertions. LTypes disappears. LStar is replaced by LAnd. This function expects its argument to be a PURE assertion.

val pp : Stdlib.Format.formatter -> t -> unit

Pretty-printer

val star : t list -> t

star [a1; a2; ...; an] will return [a1 * a2 * ... * an]

val subst_clocs : (string -> Expr.t) -> t -> t

subst_clocs subst a Substitutes expressions of the form Lit (Loc l) with subst l in a