Module Gil_syntax.Formula

type t =

| True
Logical true

| False
Logical false

| Not of t
Logical negation

| And of t * t
Logical conjunction

| Or of t * t
Logical disjunction

| Eq of Expr.t * Expr.t
Expression equality

| Less of Expr.t * Expr.t
Expression less-than for numbers

| LessEq of Expr.t * Expr.t
Expression less-than-or-equal for numbers

| StrLess of Expr.t * Expr.t
Expression less-than for strings

| SetMem of Expr.t * Expr.t
Set membership

| SetSub of Expr.t * Expr.t
Set subsetness

| ForAll of (string * Type.t option) list * t
Forall

GIL Formulae

module Set : Stdlib.Set.S with type Set.elt := t

Sets of formulae

val map : (t -> t * bool) option -> (t -> t) option -> (Expr.t -> Expr.t) option -> t -> t

Deprecated. Use Visitors.map instead

val fold : (Expr.t -> 'a) option -> (t -> 'b -> 'b -> 'a list -> 'a) -> (t -> 'b -> 'b) option -> 'b -> t -> 'a

Deprecated. Use Visitors.reduce instead

val lvars : t -> Utils.Containers.SS.t

Get all the logical variables

val pvars : t -> Utils.Containers.SS.t

Get all the program variables

val alocs : t -> Utils.Containers.SS.t

Get all the abstract locations

val clocs : t -> Utils.Containers.SS.t

Get all the concrete locations

val lists : t -> Expr.t list

Get all the logical expressions of the formula of the form (Lit (LList lst)) and (EList lst)

val list_lexprs : t -> Expr.t list

Get all the list expressions

val push_in_negations : t -> t

push_in_negations a takes negations off the toplevel of a and pushes them in the leaves. For example push_in_negations (Not (And (True, False))) returns Or (False, False)

val pp : Stdlib.Format.formatter -> t -> unit

Pretty-printer

val full_pp : Stdlib.Format.formatter -> t -> unit

Pretty-printer with constructors (will not parse)

val lift_logic_expr : Expr.t -> (t * t) option

Lifts an expression to a formula, if possible. It returns the lifted expression and its negation

val to_expr : t -> Expr.t option

Unlifts the formula to an expression, if possible

val conjunct : t list -> t

conjunct [a1; ...; an] returns a1 /\ ... /\ an

val disjunct : t list -> t

disjunct [a1; ...; an] returns a1 \/ ... \/ an

val subst_expr_for_expr : to_subst:Expr.t -> subst_with:Expr.t -> t -> t

val subst_clocs : (string -> Expr.t) -> t -> t

subst_clocs subst e Substitutes expressions of the form Lit (Loc l) with subst l in e

val get_disjuncts : t -> t list

get_disjuncts (a1 \/ ... \/ an) returns [a1; ...; an]

val strings_and_numbers : t -> string list * float list

Returns a list of strings and a list of numbers that are contained in the formula