Module Gil_syntax.Proc

type ('annot, 'label) t = {

proc_name : string;

proc_body : ('annot * 'label option * 'label Cmd.t) array;

proc_params : string list;

proc_spec : Spec.t option;

}

Labeled procedures. Every command is annotated with a label, and the gotos indicate to which label one should jump. Labels can be of any type. However, we say "labeled" when the labels are strings, and "indexed" when the labels are integers. Most functions in Gillian that work with indexed procedures assume for efficiency that the label of the i-th command is always Some i (starting from 0).

val get_params : ('a, 'b) t -> string list

Gets the parameters of the procedure

val pp : show_labels:bool -> pp_label:'a Fmt.t -> Stdlib.Format.formatter -> ('b, 'a) t -> unit

If the show_labels flag is true, the labels will be written before the command they correspond to

val pp_labeled : Stdlib.Format.formatter -> ('a, string) t -> unit

Print labelled

val pp_indexed : Stdlib.Format.formatter -> ('a, int) t -> unit

Print indexed

val line_info : (Annot.t, 'a) t -> (string * int * int) list

Line information

val indexed_of_labeled : (Annot.t, string) t -> (Annot.t, int) t

Returns the indexed procedure for a labeled procedures where the labels can be of any type. Equality of labels is decided by structural equality

val check_proc_spec_correspondence : (string, (Annot.t, 'a) t) Stdlib.Hashtbl.t -> unit