Module Exec_map.Packaged

An Exec_map to be passed to the debugger frontend and displayed

An Exec_map to be passed to the debugger frontend and displayed

type id = Logging.Report_id.t

val id_to_yojson : id -> Yojson.Safe.t

val id_of_yojson : Yojson.Safe.t -> id Ppx_deriving_yojson_runtime.error_or

type branch_case = Yojson.Safe.t

val branch_case_to_yojson : branch_case -> Yojson.Safe.t

val branch_case_of_yojson : Yojson.Safe.t -> branch_case Ppx_deriving_yojson_runtime.error_or

type cmd_data = {

id : Logging.Report_id.t;

all_ids : Logging.Report_id.t list;

display : string;

matches : Match_map.matching list;

errors : string list;

submap : id Debugger.Utils.Exec_map.submap;

}

val cmd_data_to_yojson : cmd_data -> Yojson.Safe.t

val cmd_data_of_yojson : Yojson.Safe.t -> cmd_data Ppx_deriving_yojson_runtime.error_or

type branch_data = string

val branch_data_to_yojson : branch_data -> Yojson.Safe.t

val branch_data_of_yojson : Yojson.Safe.t -> branch_data Ppx_deriving_yojson_runtime.error_or

type nonrec node = (Logging.Report_id.t, branch_case, cmd_data, branch_data) Debugger.Utils.Exec_map.node

type t = (Logging.Report_id.t, branch_case, cmd_data, branch_data) Debugger.Utils.Exec_map.map

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val _ : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

val package_gil_case : Gil_syntax.Branch_case.t -> branch_case * branch_data

Converts a GIL branch case to a packaged branch case

val package : ('i -> Logging.Report_id.t) -> (('i, 'c, 'd, 'bd) Debugger.Utils.Exec_map.node -> (Logging.Report_id.t, branch_case, cmd_data, branch_data) Debugger.Utils.Exec_map.node) -> ('i, 'c, 'd, 'bd) Debugger.Utils.Exec_map.map -> t

Converts an Exec_map to a packaged Exec_map