Module Gillian.Logging

Modules for logging (to the file log and the report database)

module Logging_constants : sig ... end

module Mode : sig ... end

Types and functions for controlling log levels

module Report_id : sig ... end

module Loggable : sig ... end

Module specifying functions required for a type to be loggable

module Log_queryer : sig ... end

Module for querying the reports stored by the database reporter. Queries will return None if the Log_queryer is not enabled.

val normal : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> ((('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a) -> unit) -> unit

Logs a message at the Normal logging level given a message format

val verbose : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> ((('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a) -> unit) -> unit

Logs a message at the Verbose logging level given a message format

val tmi : ?title:string -> ?severity:Logging.Logging_constants.Severity.t -> ((('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a) -> unit) -> unit

Logs a message at the TMI logging level given a message format

val fail : string -> 'a

Writes the string and then raises a failure.

module Specific : sig ... end

Logs a type given a Loggable.t and its content type (which should be one of the predefined strings in Logging_constants.Content_type). Returns the logged report id if it has been logged.

module Phase : sig ... end

Functions for managing phases

module Parent : sig ... end

Functions for managing the parent of new reports

val dummy_pp : Stdlib.Format.formatter -> 'a -> unit

A dummy pretty-printer

Use this when you need a pretty printer, but don't expect it to actually be seen anywhere

val print_to_all : string -> unit

Prints a message to all available reporters, and stdout if applicable.