Module Logging

module Logging_constants = Logging_constants

module Mode : sig ... end

module Report_id : sig ... end

module Reporter : sig ... end

val database_reporter : Reporter.t

val file_reporter : Reporter.t

module Loggable : module type of Loggable

module Log_queryer : sig ... end

module Report_state : sig ... end

val initialize : Reporter.t list -> unit

Initializes the logging module with the specified reporters and initializes the reporters

val wrap_up : unit -> unit

Runs any clean up code

val normal : ?title:string -> ?severity:Logging_constants.Severity.t -> ((('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a) -> unit) -> unit

val verbose : ?title:string -> ?severity:Logging_constants.Severity.t -> ((('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a) -> unit) -> unit

val tmi : ?title:string -> ?severity:Logging_constants.Severity.t -> ((('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a) -> unit) -> unit

Logs a message at the TMI logging level given a message format

module Specific : sig ... end

val fail : string -> 'a

Writes the string and then raises a failure.

val print_to_all : string -> unit

Output the strings in every file and prints it to stdout

module Phase : sig ... end

val set_previous : ?force_none:bool -> Gillian.Logging.Report_id.t option -> unit

module Parent : sig ... end

val dummy_pp : Stdlib.Format.formatter -> 'a -> unit