Module ParserAndCompiler

module type S = sig ... end

This defines an interface that allows a user to indicate how to parse their own programming language, preprocess the obtained language and compile it to GIL (type LabProg.t)

module Dummy : S

Dummy ParserAndCompiler that will simply always fail. This is used when someone wants to build a command line interface to only reason about GIL.